minor graph middleware refactor, no logic changes
